Create README.md
This code was written for a Visual Basic application that calculates monthly loan payments over a period of years. On the first form (ClientData), the client enters their name and clicks a button, which opens another form (PaymentEstimator) allowing them to fill out their basic loan data. On this form, the client then enters the principal value and interest rate of their loan, selects a term of 1, 10, or 30 years, and clicks a button which displays the amount of the monthly payment for the given loan data they have entered next to the interest rate.